Worked on Level | Organize | Optimize Wolf/FIGHT-Wall
Worked on Level:
Added the next Arena-Part of the Level and made some platforms easier to
reach

Organize:
Creatzed folder to seperate each tile of the landscape and the
characters for easier changes in the level

Optimize Wolf:
Made a boolean pin called "just rotate". If this is set on (you can
select it in the Editor) the Wolf dont use the walk animation and now
just rotates.

Optimize FIGHT-Wall:
Made the Amount of kills you need to reach changeable in the Editor, so
you dont need to recreate the wall as bluebprint if you need 5 instead
of 2 kills.